requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
babel==2.16.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==5.1.1
django-cors-headers==4.4.0
django-money==3.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
fonttools==4.54.1
fpdf2==2.8.1
idna==3.9
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jalali_core==1.0.0
jdatetime==5.0.0
kavenegar==1.1.2
kombu==5.4.2
Naked==0.1.32
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
# psycopg2==2.9.10
psycopg2-binary==2.9.10
py-moneyed==3.0
pycryptodome==3.21.0
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
sentry-sdk==2.17.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.0
wcwidth==0.2.13
